paparece	07/06/2018 - Media.

+++ system/application/controllers/maint.php
@@ -234,10 +234,10 @@ class Maint extends CI_Controller {
     try{
         $usb_path = $this->config->item('usb_path');
         $good_usb_path = "";
-        $sdavolcmd = "/usr/bin/sudo /sbin/fdisk -l | /bin/grep /dev/sd | /bin/grep exFAT | /bin/grep NTFS | /bin/grep HPFS | /usr/bin/awk -F ' ' '{/usr/bin/print $1}'";
+        $sdavolcmd = "/usr/bin/sudo /sbin/fdisk -l | /bin/grep /dev/sd | /bin/grep exFAT | /bin/grep NTFS | /bin/grep HPFS | /usr/bin/awk -F ' ' '{print $1}'";
         $sdavol = exec($sdavolcmd);
         if (strlen($sdavol) > 0){
-            $chkvolcmd = "/bin/cat /proc/mounts | /bin/grep '.$usb_path.'| /bin/grep '.$sdavol.' | /usr/bin/awk -F ' ' '{/usr/bin/print $2}'";
+            $chkvolcmd = "/bin/cat /proc/mounts | /bin/grep '.$usb_path.'| /bin/grep '.$sdavol.' | /usr/bin/awk -F ' ' '{print $2}'";
             $volname = exec($chkvolcmd);
             if (strlen($chkvolcmd) === 0){
                 $mntcmd = '/usr/bin/sudo /bin/mount -o uid-www-data,gid-www-data '.$sdavol.' /media/pidrive';
@@ -248,7 +248,7 @@ class Maint extends CI_Controller {
         if (file_exists($usb_path)){
             $directories = glob($usb_path . '*' , GLOB_ONLYDIR);
             foreach($directories as $directory){
-                if((is_writable($directory) == TRUE) && (is_dir_empty($directory) == FALSE)){
+                if((is_writable($directory) == TRUE) && (count($filelist) > 2)){ //&& (is_dir_empty($directory) == FALSE)){
                     if(touch($directory.'/test.txt') == TRUE){
                         if(unlink($directory.'/test.txt') == TRUE){
                             $good_usb_path = $directory;
